# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NUTTER-XMD SaaS platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type BotStatus = (typeof BotStatus)[keyof typeof BotStatus];

export const BotStatus = {
  offline: "offline",
  connecting: "connecting",
  online: "online",
  banned: "banned",
} as const;

export interface Bot {
  id: number;
  userId: string;
  name: string;
  /** @nullable */
  description?: string | null;
  status: BotStatus;
  isActive: boolean;
  /** @nullable */
  phoneNumber?: string | null;
  /** @nullable */
  sessionData?: string | null;
  autoReply: boolean;
  /** @nullable */
  autoReplyMessage?: string | null;
  prefix: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateBotBody {
  name: string;
  /** @nullable */
  description?: string | null;
  prefix: string;
  autoReply?: boolean;
  /** @nullable */
  autoReplyMessage?: string | null;
}

export interface UpdateBotBody {
  name?: string;
  /** @nullable */
  description?: string | null;
  prefix?: string;
  autoReply?: boolean;
  /** @nullable */
  autoReplyMessage?: string | null;
  isActive?: boolean;
}

export type QRCodeResponseStatus =
  (typeof QRCodeResponseStatus)[keyof typeof QRCodeResponseStatus];

export const QRCodeResponseStatus = {
  pending: "pending",
  ready: "ready",
  connected: "connected",
  error: "error",
} as const;

export interface QRCodeResponse {
  /** @nullable */
  qrCode?: string | null;
  status: QRCodeResponseStatus;
}

export interface BotCommand {
  id: number;
  botId: number;
  command: string;
  /** @nullable */
  description?: string | null;
  response: string;
  isEnabled: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface CreateBotCommandBody {
  command: string;
  /** @nullable */
  description?: string | null;
  response: string;
  isEnabled?: boolean;
}

export interface UpdateBotCommandBody {
  command?: string;
  /** @nullable */
  description?: string | null;
  response?: string;
  isEnabled?: boolean;
}

export interface DashboardStats {
  totalBots: number;
  activeBots: number;
  onlineBots: number;
  totalCommands: number;
}

export interface AdminLoginBody {
  username: string;
  key: string;
}

export interface AdminLoginResponse {
  token: string;
  message: string;
}

export interface AdminBot {
  id: number;
  userId: string;
  name: string;
  status: string;
  isActive: boolean;
  /** @nullable */
  phoneNumber?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface AdminStats {
  totalUsers: number;
  totalBots: number;
  activeBots: number;
  onlineBots: number;
  totalCommands: number;
}