# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NUTTER-XMD SaaS platform API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all bots for authenticated user
 */
export const ListBotsResponseItem = zod.object({
  id: zod.number(),
  userId: zod.string(),
  name: zod.string(),
  description: zod.string().nullish(),
  status: zod.enum(["offline", "connecting", "online", "banned"]),
  isActive: zod.boolean(),
  phoneNumber: zod.string().nullish(),
  sessionData: zod.string().nullish(),
  autoReply: zod.boolean(),
  autoReplyMessage: zod.string().nullish(),
  prefix: zod.string(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const ListBotsResponse = zod.array(ListBotsResponseItem);

/**
 * @summary Create a new bot
 */
export const CreateBotBody = zod.object({
  name: zod.string(),
  description: zod.string().nullish(),
  prefix: zod.string(),
  autoReply: zod.boolean().optional(),
  autoReplyMessage: zod.string().nullish(),
});

/**
 * @summary Get a bot by ID
 */
export const GetBotParams = zod.object({
  id: zod.coerce.number(),
});

export const GetBotResponse = zod.object({
  id: zod.number(),
  userId: zod.string(),
  name: zod.string(),
  description: zod.string().nullish(),
  status: zod.enum(["offline", "connecting", "online", "banned"]),
  isActive: zod.boolean(),
  phoneNumber: zod.string().nullish(),
  sessionData: zod.string().nullish(),
  autoReply: zod.boolean(),
  autoReplyMessage: zod.string().nullish(),
  prefix: zod.string(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Update a bot
 */
export const UpdateBotParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateBotBody = zod.object({
  name: zod.string().optional(),
  description: zod.string().nullish(),
  prefix: zod.string().optional(),
  autoReply: zod.boolean().optional(),
  autoReplyMessage: zod.string().nullish(),
  isActive: zod.boolean().optional(),
});

export const UpdateBotResponse = zod.object({
  id: zod.number(),
  userId: zod.string(),
  name: zod.string(),
  description: zod.string().nullish(),
  status: zod.enum(["offline", "connecting", "online", "banned"]),
  isActive: zod.boolean(),
  phoneNumber: zod.string().nullish(),
  sessionData: zod.string().nullish(),
  autoReply: zod.boolean(),
  autoReplyMessage: zod.string().nullish(),
  prefix: zod.string(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Delete a bot
 */
export const DeleteBotParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get QR code for bot WhatsApp connection
 */
export const GetBotQRParams = zod.object({
  id: zod.coerce.number(),
});

export const GetBotQRResponse = zod.object({
  qrCode: zod.string().nullish(),
  status: zod.enum(["pending", "ready", "connected", "error"]),
});

/**
 * @summary Disconnect bot from WhatsApp
 */
export const DisconnectBotParams = zod.object({
  id: zod.coerce.number(),
});

export const DisconnectBotResponse = zod.object({
  id: zod.number(),
  userId: zod.string(),
  name: zod.string(),
  description: zod.string().nullish(),
  status: zod.enum(["offline", "connecting", "online", "banned"]),
  isActive: zod.boolean(),
  phoneNumber: zod.string().nullish(),
  sessionData: zod.string().nullish(),
  autoReply: zod.boolean(),
  autoReplyMessage: zod.string().nullish(),
  prefix: zod.string(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary List commands for a bot
 */
export const ListBotCommandsParams = zod.object({
  id: zod.coerce.number(),
});

export const ListBotCommandsResponseItem = zod.object({
  id: zod.number(),
  botId: zod.number(),
  command: zod.string(),
  description: zod.string().nullish(),
  response: zod.string(),
  isEnabled: zod.boolean(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const ListBotCommandsResponse = zod.array(ListBotCommandsResponseItem);

/**
 * @summary Create a command for a bot
 */
export const CreateBotCommandParams = zod.object({
  id: zod.coerce.number(),
});

export const CreateBotCommandBody = zod.object({
  command: zod.string(),
  description: zod.string().nullish(),
  response: zod.string(),
  isEnabled: zod.boolean().optional(),
});

/**
 * @summary Update a bot command
 */
export const UpdateBotCommandParams = zod.object({
  id: zod.coerce.number(),
  commandId: zod.coerce.number(),
});

export const UpdateBotCommandBody = zod.object({
  command: zod.string().optional(),
  description: zod.string().nullish(),
  response: zod.string().optional(),
  isEnabled: zod.boolean().optional(),
});

export const UpdateBotCommandResponse = zod.object({
  id: zod.number(),
  botId: zod.number(),
  command: zod.string(),
  description: zod.string().nullish(),
  response: zod.string(),
  isEnabled: zod.boolean(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Delete a bot command
 */
export const DeleteBotCommandParams = zod.object({
  id: zod.coerce.number(),
  commandId: zod.coerce.number(),
});

/**
 * @summary Get dashboard statistics for the current user
 */
export const GetDashboardStatsResponse = zod.object({
  totalBots: zod.number(),
  activeBots: zod.number(),
  onlineBots: zod.number(),
  totalCommands: zod.number(),
});

/**
 * @summary Admin login
 */
export const AdminLoginBody = zod.object({
  username: zod.string(),
  key: zod.string(),
});

export const AdminLoginResponse = zod.object({
  token: zod.string(),
  message: zod.string(),
});

/**
 * @summary Admin - list all bots across all users
 */
export const AdminListAllBotsResponseItem = zod.object({
  id: zod.number(),
  userId: zod.string(),
  name: zod.string(),
  status: zod.string(),
  isActive: zod.boolean(),
  phoneNumber: zod.string().nullish(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const AdminListAllBotsResponse = zod.array(AdminListAllBotsResponseItem);

/**
 * @summary Admin - activate a bot
 */
export const AdminActivateBotParams = zod.object({
  id: zod.coerce.number(),
});

export const AdminActivateBotResponse = zod.object({
  id: zod.number(),
  userId: zod.string(),
  name: zod.string(),
  status: zod.string(),
  isActive: zod.boolean(),
  phoneNumber: zod.string().nullish(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Admin - deactivate a bot
 */
export const AdminDeactivateBotParams = zod.object({
  id: zod.coerce.number(),
});

export const AdminDeactivateBotResponse = zod.object({
  id: zod.number(),
  userId: zod.string(),
  name: zod.string(),
  status: zod.string(),
  isActive: zod.boolean(),
  phoneNumber: zod.string().nullish(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Admin - platform-wide statistics
 */
export const AdminGetStatsResponse = zod.object({
  totalUsers: zod.number(),
  totalBots: zod.number(),
  activeBots: zod.number(),
  onlineBots: zod.number(),
  totalCommands: zod.number(),
});